LayoutTests/fullscreen/full-screen-enter-while-exiting.html

<!DOCTYPE html>
<html>
<head>
    <title>full-screen-enter-while-exiting</title>
    <script src="full-screen-test.js"></script>
    <script>

    window.addEventListener('load', async event => {
        if (Element.prototype.webkitRequestFullScreen == undefined) {
            logResult(false, "Element.prototype.webkitRequestFullScreen == undefined");
            endTest();
            return;
        }

        if (!window.testRunner || !window.testRunner.isWebKit2) {
            logResult(false, "This test requires WebKitTestRunner.");
            endTest();
            return;
        }

        testRunner.waitBeforeFinishingFullscreenExit();

        consoleWrite('Enter fullscreen with target1');
        internals.withUserGesture(() => { run('target1.webkitRequestFullScreen()'); });
        await waitFor(target1, 'webkitfullscreenchange');
        await testRunner.updatePresentation();

        consoleWrite('Exit fullscreen with target1');
        run('document.webkitExitFullscreen()');
        changePromise = waitFor(document, 'webkitfullscreenchange')
        await testRunner.updatePresentation();

        consoleWrite('Attempt to enter fullscreen with target2');
        internals.withUserGesture(() => { run('target2.webkitRequestFullScreen()'); });

        await waitFor(target2, 'webkitfullscreenerror');

        testRunner.finishFullscreenExit();
        await testRunner.updatePresentation();

        internals.withUserGesture(() => { run('target2.webkitRequestFullScreen()'); });
        await waitFor(target2, 'webkitfullscreenchange');

        endTest();
    });
</script>
</head>
<body>
<span id="target1"></span>
<span id="target2"></span>
</body>
</html>
