LayoutTests/fullscreen/full-screen-enter-while-exiting-expected.txt

Enter fullscreen with target1
RUN(target1.webkitRequestFullScreen())
EVENT(webkitfullscreenchange)
Exit fullscreen with target1
RUN(document.webkitExitFullscreen())
Attempt to enter fullscreen with target2
RUN(target2.webkitRequestFullScreen())
EVENT(webkitfullscreenerror)
EVENT(webkitfullscreenchange)
RUN(target2.webkitRequestFullScreen())
EVENT(webkitfullscreenchange)
END OF TEST
